+++ R/proc_sort.R
@@ -20,6 +20,28 @@
 #' the by parameter if passed.  Otherwise, the function will dedupe on
 #' all variables returned.}
 #' }
+#' @section Missing Values:
+#' Missing values are handled in R differently than they are handled in SAS.
+#' In R, missing values (NA) are sorted last, and most sorting
+#' functions in R give you an option to sort it first if desired.
+#'
+#' SAS, on the other hand, considers a missing as the smallest value.  If the
+#' sort is ascending, it will sort the missing values first.  If the sort is
+#' descending, it will sort the missing values last.
+#'
+#' The \code{na.sort} parameter on \code{proc_sort} gives you a choice of
+#' how NAs are sorted.  By default, the function sorts in the standard R
+#' manner.  If you are trying to replicate a SAS sort, set the
+#' value to "sas".
+#'
+#' Note that this parameter only sorts NA values in a manner similar to SAS.
+#' It does not guarantee that the entire sort will match SAS.  There can
+#' still be differences in the sort caused by formats, character set mismatch,
+#' or other reasons.
+#'
+#' If you want to always use the SAS style sort, you can set it as a global
+#' option, like this: \code{options("procs.na.sort" = "sas")}.
+#'
 #' @param data The input data to sort.
 #' @param by A vector of variables to sort by.
 #' @param keep A vector of variables on the output data to keep.  All other
@@ -39,6 +61,15 @@
 #' variable as a factor.  This parameter therefore allows you to use
 #' the factor for the sort, but then convert back to a character
 #' once the sort is complete.
+#' @param na.sort An option that determines how to sort NA values.  Valid
+#' values are NULL, "first", "last", and "sas". Values may be quoted or
+#' unquoted. The "sas"
+#' value will sort NAs first for ascending keys, and last for descending
+#' keys.  The NA sort style may be controlled globally with
+#' \code{options("procs.na.sort" = <value>)}.  Any values passed
+#' on the local parameter will override the global setting. The default
+#' value is NULL, which defers to the global setting.  If the global setting
+#' is NULL, the parameter will default to "last".
 #' @return The sorted dataset.  If a data frame was input, a
 #' data frame will be output.  If a tibble was input, a tibble will
 #' be output.
@@ -105,7 +136,7 @@
 #' @import tibble
 #' @export
 proc_sort <- function(data,  by = NULL, keep = NULL, order = "ascending",
-                      options = NULL, as.character = FALSE) {
+                      options = NULL, as.character = FALSE, na.sort = NULL) {
 
 
   # Deal with single value unquoted parameter values
@@ -128,6 +159,16 @@ proc_sort <- function(data,  by = NULL, keep = NULL, order = "ascending",
   oopt <- deparse(substitute(options, env = environment()))
   options <- tryCatch({if (typeof(options) %in% c("integer", "double", "character", "NULL")) options else oopt},
                       error = function(cond) {oopt})
+
+  # Deal with single value unquoted option values
+  onasort <- deparse(substitute(na.sort, env = environment()))
+  na.sort <- tryCatch({if (typeof(na.sort) %in% c("character", "NULL")) na.sort else onasort},
+                      error = function(cond) {onasort})
+
+  # Force to lower case
+  if (!is.null(na.sort)) {
+    na.sort <- tolower(na.sort)
+  }
 
   # Parameter checks
 
@@ -171,8 +212,34 @@ proc_sort <- function(data,  by = NULL, keep = NULL, order = "ascending",
 
   }
 
+  if (is.null(na.sort)) {
+    # Get any global sort settings
+    osrt <- options("procs.na.sort")
+    if (!is.null(osrt$procs.na.sort)) {
+      na.sort <- osrt$procs.na.sort
+    } else {
+      na.sort <- "last"
+    }
+  }
 
-  ret <- sort(data, by = by, ascending = asc )
+  nsrt <- TRUE
+  if (na.sort == "first") {
+    nsrt <- FALSE
+  } else if (na.sort == "last") {
+    nsrt <- TRUE
+  } else if (na.sort == "sas") {
+    nsrt <- c()
+    for (av in asc) {
+      if (av) {
+        nsrt <- append(nsrt, FALSE)
+      } else {
+        nsrt <- append(nsrt, TRUE)
+      }
+    }
+  }
+
+  # Perform sort
+  ret <- sort(data, by = by, ascending = asc, na.last = nsrt)
 
   # Order and keep
   ret <- ret[ , keep, drop = FALSE]