+++ NEWS.md
@@ -7,6 +7,8 @@
 * Added "AIC" statistic to `proc_reg()`.
 * Added "weight" parameter to `proc_means()`.
 * Added "vardef" option to `proc_means()`.
+* Added "na.sort" parameter to `proc_sort()`.
+* Updates to documentation.
 
 # procs 1.0.7